+++ packages/sdk-ink/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+### Added
+
+- Utilities
+  - `getDeploymentAddressWithNonce` to estimate a revive deployment address with an account nonce (equivalent to `create1` in polkadot-sdk)
+  - `getDeploymentAddressWithSalt` to estimate a revive deployment address with salt (equivalent to `create2` in polkadot-sdk)
+  - `getDeploymentHash` to calculate the deployment hash for a particular code + constructor call (needed for `getDeploymentAddressWithSalt`)
+
 ### Fixed
 
 - decode events when dry-running in pallet revive.